refactor(tools/get_unique_path): extract ScrapperObjectCore.get_unique_path to tools/get_unique_path
